developed api fetch logic and displayed 7 day data in main component
