+++ mera-weather-app/src/hooks/use-http.js
@@ -22,8 +22,9 @@ const useHttp = () => {
       applyData(data);
     } catch (err) {
       setError(err.message || "Something went wrong!");
+    } finally {
+      setIsLoading(false);
     }
-    setIsLoading(false);
   }, []);
 
   return {
