--- mera-weather-app/src/util/fetchFunction.js
@@ -1,10 +1,0 @@
-export const fetchFunction = async (url, applyData) => {
-  try {
-    const response = await fetch(url);
-    const json = await response.json();
-    applyData(json);
-  } catch (err) {
-    throw err;
-    console.log(err);
-  }
-};
